Split components into files
